+++ www/api_proxy.php
@@ -141,6 +141,7 @@ $ALLOWED_PROXY_PREFIXES = [
     '/policy/',
     '/drift/',
     '/tasks/',
+    '/groups',
 ];
 $pathAllowed = false;
 foreach ($ALLOWED_PROXY_PREFIXES as $prefix) {
@@ -168,6 +169,7 @@ $ADMIN_ONLY_PREFIXES = [
     '/scan_server_users', '/sshd_allow_user', '/remove_user_keys', '/delete_remote_user',
     '/server_user_remove_key', '/admin/', '/policy/', '/exclude_user',
     '/server_lifecycle', '/update_security_exec', '/drift/', '/tasks/',
+    '/groups',
 ];
 if ($roleId < ROLE_ADMIN) {
     foreach ($ADMIN_ONLY_PREFIXES as $prefix) {